+++ cogs/timetracker.py
@@ -110,20 +110,6 @@ class timeTracker(commands.Cog):
 
                 await ctx.response.send_message(f"{ctx.user.display_name} has clocked out adding {time_added}.")
 
-    async def check_manager_perms(ctx):
-            # config import
-            f = open('.config')
-            config = json.load(f)
-            admin_role = config["permissions"]["admin"]
-            print(F" the admin role is '{admin_role}'")
-            role = discord.utils.get(ctx.guild.roles, name=admin_role)
-            if role in ctx.user.roles:
-                return True
-            elif role not in ctx.user.roles:
-                await ctx.response.send_message("Sorry you do not have the correct permissions.", ephemeral=True)
-            else:
-                await ctx.response.send_message("Sorry something went wrong.", ephemeral=True)
-
     @discord.app_commands.command(name="get_user_time", description="get time entries of a specific user. RESTRICTED")
     @discord.app_commands.check(check_manager_perms)
     async def get_user_time(self, ctx, user: discord.Member, csv_output: bool = False):